fix: update fetchCredentialStatusVC to use documentLoader and add cache for jws 2020 v1
