+++ packages/w3c-vc/src/index.ts
@@ -1,14 +1,20 @@
 import { verifyCredentialStatus } from './lib/verify/credentialStatus';
 import {
   deriveCredential,
-  getDocumentLoader,
   isRawDocument,
   isSignedDocument,
   signCredential,
   verifyCredential,
 } from './lib/w3c-vc';
+import { getDocumentLoader } from '@trustvc/w3c-context';
 
 export * from './lib/types';
+export {
+  ContextDocument,
+  DocumentLoader,
+  Document,
+  DocumentLoaderObject,
+} from '@trustvc/w3c-context';
 export { CredentialStatusResult } from './lib/verify/credentialStatus/types';
 export {
   deriveCredential,
